Change CorrelationId from int to long

+++ API/src/TibiaEnemyOtherCharactersFinder.Domain/Entities/CharacterCorrelation.cs
@@ -5,8 +5,7 @@ public class CharacterCorrelation : IEntity
     /// <summary>
     /// Id of specific correlation between two characters
     /// </summary>
-    public int CorrelationId { get; set; }
-    // UNDONE: change int to long
+    public long CorrelationId { get; set; }
 
     /// <summary>
     /// Id of specific character that logout
